backend/pyproject.toml
[tool.poetry]
name = "strickland-cannon-coop"
version = "0.1.0"
description = ""
authors = ["Joshua Cannon <joshdcannon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
python-dateutil = "^2.8.1"
docopt = "^0.6.2"
requests = "^2.24.0"
tqdm = "^4.47.0"
selenium = "^3.141.0"
scipy = "^1.5.1"
python-gnupg = "^0.4.6"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
